tests/wpt/web-platform-tests/presentation-api/controlling-ua/startNewPresentation_sandboxing_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Sandboxing: starting a presentation from a nested context succeeds when allow-presentation is set</title>
<link rel="author" title="Francois Daoust" href="https://www.w3.org/People/#fd">
<link rel="help" href="http://w3c.github.io/presentation-api/#dom-presentationrequest-start">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe id="childFrame" sandbox="allow-scripts allow-presentation" style="display:none" src="support/iframe.html"></iframe>
<p>Click the button below to start the manual test. If prompted to select a device, please dismiss the dialog box. The test passes if a "PASS" result appears.</p>
<button id="presentBtn" onclick="startPresentationTest()">Start Presentation Test</button>

<script>
    setup({explicit_timeout: true});

    var presentBtn = document.getElementById('presentBtn');
    var childFrame = document.getElementById('childFrame');

    function startPresentationTest() {
        presentBtn.disabled = true;
        async_test(function (t) {
            childFrame.contentWindow.postMessage('start', '*');
            window.addEventListener('message', t.step_func(function (ev) {
                assert_equals(ev.data, 'success',
                    'Presentation could not be started from nested frame.');
                t.done();
            }));
        });
    }
</script>
